refactor: WP-876 move tenant UI config
